% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{cellular.fraction}
\alias{cellular.fraction}
\title{helper function for app}
\usage{
cellular.fraction(out, fit, method = c("cncf", "em"),
  load.genome = FALSE, gene.pos = NULL, main = "", lend = "butt",
  theme = "bw", ...)
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
