% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{close.up}
\alias{close.up}
\title{helper function for app}
\usage{
\method{close}{up}(out, fit, chrom.range = NULL, method = NA,
  gene.name = NULL, lend = "butt", bed.path = NULL,
  cached.gene.path = NULL, subset.snps = FALSE, ...)
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
