% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{copy.number.log.ratio}
\alias{copy.number.log.ratio}
\title{helper function for app}
\usage{
copy.number.log.ratio(out, fit, load.genome = FALSE, gene.pos = NULL,
  col.1 = "#0080FF", col.2 = "#4CC4FF", sample.num = NULL,
  lend = "butt", theme = "bw", subset.indices = NULL)
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
