% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get.cumulative.chr.maploc}
\alias{get.cumulative.chr.maploc}
\title{helper function for app}
\usage{
get.cumulative.chr.maploc(mat, load.genome = FALSE)
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
