% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get.gene.pos.cached}
\alias{get.gene.pos.cached}
\title{helper function for app}
\usage{
get.gene.pos.cached(hugo.symbol,
  my.path = "/ifs/work/bandlamc/git/facets-suite/Homo_sapiens.GRCh37.75.gene_positions.txt")
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
