% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_review_status}
\alias{get_review_status}
\title{helper function for app}
\usage{
get_review_status(sample, sample_path)
}
\arguments{
\item{sample}{sampleid}

\item{sample_path}{facets run directory containing 'facets_review.manifest'}
}
\value{
converts string to numeric and rounds to 2-digits
}
\description{
helper function for app
}
