% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{integer.copy.number}
\alias{integer.copy.number}
\title{helper function for app}
\usage{
integer.copy.number(out, fit, method = c("cncf", "em"),
  load.genome = FALSE, gene.pos = NULL, main = "", lend = "butt",
  theme = "bw", ...)
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
