% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{load_samples}
\alias{load_samples}
\title{helper function for app}
\usage{
load_samples(manifest, progress)
}
\arguments{
\item{manifest}{list of facets run directories}

\item{progress}{progress bar from shiny}
}
\value{
simple metadata data.frame
}
\description{
helper function for app
}
