% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{metadata_init}
\alias{metadata_init}
\title{helper function for app}
\usage{
metadata_init(sample, sample_path, progress)
}
\arguments{
\item{sample}{sample name}

\item{sample_path}{path to facets run dir}

\item{progress}{progress bar from shiny}
}
\value{
description of the facets run
}
\description{
helper function for app
}
