% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{metadata_init_quick}
\alias{metadata_init_quick}
\title{helper function for app}
\usage{
metadata_init_quick(sample, sample_path)
}
\arguments{
\item{sample}{sample name}

\item{sample_path}{path to facets run dir}
}
\value{
minimal description of the facets run
}
\description{
helper function for app
}
