% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{round_down}
\alias{round_down}
\title{helper function for app}
\usage{
round_down(val_str)
}
\arguments{
\item{val_str}{input string}
}
\value{
converts string to numeric and rounds to 2-digits
}
\description{
helper function for app
}
