% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{update_review_status_file}
\alias{update_review_status_file}
\title{helper function for app}
\usage{
update_review_status_file(sample_path, df)
}
\arguments{
\item{sample_path}{facets run directory containing 'facets_review.manifest'}

\item{df}{dataframe}
}
\value{
converts string to numeric and rounds to 2-digits
}
\description{
helper function for app
}
